LayoutTests/editing/pasteboard/copy-paste-text-with-trailing-colon-expected.txt

This test verifies that copying the text 'Foo:' and pasting into a plain text field works as expected. To run the test manually, click the button, focus the text field, and paste.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS doneWithPaste became true
PASS target.value is "Foo:"
PASS successfullyParsed is true

TEST COMPLETE
Click me
